reorganizing structure
