+++ models/__init__.py
@@ -1,0 +1,4 @@
+from .bnn import *
+from .cnn import *
+from .hypernet import *
+from .mlp import *
