+++ templates/__init__.py
@@ -1,0 +1,2 @@
+from .sgd_template import *
+from .vi_template import *
